+++ src/main/java/com/appraisaltool/model/Team.java
@@ -6,7 +6,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-
 import lombok.Data;
 
 @Data
@@ -16,7 +15,7 @@ public class Team {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	Long teamId;
+    Integer teamId;
 	
 	@Column(name = "teamName")
 	String teamName;
