+++ src/main/java/com/appraisaltool/repository/OfficeRepository.java
@@ -3,12 +3,11 @@ package com.appraisaltool.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.appraisaltool.model.Office;
 
 
 @Repository
-public interface OfficeRepository extends JpaRepository<Office, Long>{
+public interface OfficeRepository extends JpaRepository<Office, Integer> {
 	
 	 Office getOfficeByOfficeName(String officeName);
